requirements.txt
Django==1.8.3
django-extensions==1.5.5
django-filter==0.10.0
django-reversion==1.8.7
djangorestframework==3.1.3
psycopg2==2.6.1
python-dateutil==2.4.2
python-mimeparse==0.1.4
pytz==2015.4
requests==2.7.0
six==1.9.0
ua-parser==0.3.6
user-agents==0.3.2
